update structuar of project

+++ school hub/Models/Sections/Section.cs
@@ -5,7 +5,7 @@ namespace school_hub.Models
 {
     public abstract class Section : DisplayInfo
     {
-        public short SectionId { get; set; }
+        public int SectionId { get; set; }
         public enSectionType SectionType { get; set; }
     }
     public enum enSectionType
